\name{HU}
\alias{HU}

\title{Human placenta versus the Universal reference hybridization data}

\usage{data(HU)}
\description{
  This is a data frame consists 5 variables.
}
\format{
  A data frame with 75714 observations on 5 variables:
  \tabular{rll}{
  [,1] \tab HUid \tab gene identification number \cr
  [,2] \tab HUslide \tab slide number \cr
  [,3] \tab HUblock \tab block number within each slide \cr
  [,4] \tab HUint   \tab total intensity \eqn{0.5log_2(RG)} \cr
  [,5] \tab HUratio \tab intensity ratios \eqn{log_2\frac{R}{G}}\cr
  }
where R is background adjusted signal intensity in the Cy5 channel,
G is background adjusted signal intensity in the Cy3 channel.
}
\source{
These data is the part of data set generated by human placenta versus the universal reference 
hybridization by Dr. Bento M. Soares lab in the University of Iowa. The data sets were used 
by Wang et al. (2006,2007) for the purpose of methodology demonstration.
}
\references{
   Wang, D., Huang, J., Xie, H., Manzella, L., Soares, M. B.,  \bold{A robust two-way semi-linear model for
 normalization of cDNA microarray data},\emph{BMC Bioinformatics 2005, 6:14}.

  Wang, D., Zhang, C-H,Soares, M. B., Huang, J.(2007) \bold{Systematic approaches for 
  incorporating control spots and data quality information to improve 
  normalization of cDNA microarray data}, \emph{Journal of Biopharmaceutical 
  Statistics, 17:415-431}.
}
\keyword{datasets}


