\name{HUweightdata}
\alias{HUweightdata}

\title{Spot quality information for Human placenta versus the Universal reference  hybridization}

\usage{data(HUweightdata)}
\description{
  This data set consists 10 variables.
}
\format{
   A list with 75714 observations on 10 variables:
  \tabular{rll}{
   [,1] \tab  RF \tab front mean intensity for the Cy5 \cr
   [,2] \tab RB \tab background mean intensity for the Cy5 \cr
   [,3] \tab   SDRF \tab standard deviation of RF \cr
   [,4] \tab   SDRB \tab standard deviation of RB \cr
   [,5] \tab   GF   \tab front mean intensity for the Cy3 \cr
   [,6] \tab   GB   \tab background mean intensity for the Cy3 \cr
   [,7] \tab   SDGF \tab standard deviation of GF \cr
   [,8] \tab   SDGB \tab standard deviation of GB \cr
   [,9] \tab   NF   \tab number of pixels used for the front intensity \cr
   [,10] \tab   NB   \tab number of pixels used for the background intensity \cr
    }
}
\source{
These data is the part of data set generated by human placenta versus the universal reference 
hybridization by Dr. Bento M. Soares lab in the University of Iowa. The data sets were used 
by Wang et al. (2006,2007) for the purpose of methodology demonstration.
}
\references{
   Wang, D., Huang, J., Xie, H., Manzella, L., Soares, M. B.,  \bold{A robust two-way semi-linear model for
 normalization of cDNA microarray data},\emph{BMC Bioinformatics 2005, 6:14}.

  Wang, D., Zhang, C-H,Soares, M. B., Huang, J.(2007) \bold{Systematic approaches for 
  incorporating control spots and data quality information to improve 
  normalization of cDNA microarray data}, \emph{Journal of Biopharmaceutical 
  Statistics, 17:415-431}.
}
\keyword{datasets}


